Add ErrorPromoter module to handle promote error logic
