+++ lib/yaaf/form.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
+require_relative 'error_promoter'
+
 module YAAF
   # Parent class for form objects
   class Form
     include ::ActiveModel::Model
     include ::ActiveModel::Validations::Callbacks
     include ::ActiveRecord::Transactions
+    include ErrorPromoter
     define_model_callbacks :save
 
     validate :validate_models
@@ -30,26 +33,6 @@ module YAAF
 
     attr_accessor :models
 
-    def promote_errors(model)
-      if rails_version_less_than_6_1?
-        promote_legacy_errors(model)
-      else
-        model.errors.each do |model_error|
-          errors.add(model_error.attribute, model_error.type, **model_error.options)
-        end
-      end
-    end
-
-    def promote_legacy_errors(model)
-      model.errors.details.each do |attribute, details|
-        details.each do |detail|
-          options = detail.except(:error)
-
-          errors.add(attribute, detail[:error], **options)
-        end
-      end
-    end
-
     def save_in_transaction(options)
       ::ActiveRecord::Base.transaction do
         run_callbacks :save do
@@ -73,11 +56,6 @@ module YAAF
     def handle_transaction_rollback(exception)
       run_callbacks :rollback
       raise exception
-    end
-
-    def rails_version_less_than_6_1?
-      ActiveModel::VERSION::MAJOR < 6 ||
-        ActiveModel::VERSION::MAJOR == 6 && ActiveModel::VERSION::MINOR.zero?
     end
   end
 end
